feat: Horse Detection & Bounding Box (#3)
- YOLO 기반 TFLite 모델 적용하여 말(horse) 객체 검출 로직 추가
- 검출된 Bounding Box 목록에 NMS(Non-Max Suppression) 도입
- Bounding Box 오버레이 구현
