+++ app/src/main/java/com/example/horsetracker/presentation/detector/BoundingBox.kt
@@ -1,0 +1,15 @@
+package com.example.horsetracker.presentation.detector
+
+data class BoundingBox(
+    val x1: Float,
+    val y1: Float,
+    val x2: Float,
+    val y2: Float,
+    val cx: Float,
+    val cy: Float,
+    val w: Float,
+    val h: Float,
+    val cnf: Float,
+    val cls: Int,
+    val clsName: String
+)
