+++ app/src/main/java/com/example/horsetracker/presentation/ui/ai/ModelImage.kt
@@ -1,0 +1,13 @@
+package com.example.horsetracker.presentation.ui.ai
+
+import android.graphics.Bitmap
+import androidx.compose.foundation.Image
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.asImageBitmap
+
+@Composable
+fun ModelImage(bitmap: Bitmap?) {
+    bitmap?.let {
+        Image(bitmap = it.asImageBitmap(), contentDescription = "Description of the image")
+    }
+}
